TweetScraper/requirements.txt
absl-py==0.9.0
alembic==1.2.1
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.2
bert==2.2.0
bert-tensorflow==1.0.1
bleach==1.5.0
bson==0.5.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
config==0.4.2
configparser==4.0.2
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
dataclasses==0.6
Django==2.2.9
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.0
erlastic==2.0.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mongoengine==0.19.1
mysql-connector==2.2.9
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
parsel==1.5.2
pickle-mixin==1.0.2
protobuf==3.11.2
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.10.1
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
Scrapy==1.7.3
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.8
sqlparse==0.2.4
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.7.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tweet-preprocessor==0.5.0
Twisted==19.7.0
urllib3==1.25.7
w3lib==1.21.0
Werkzeug==0.16.0
wrapt==1.11.2
zope.interface==4.6.0
